// src/Enums.ts
export enum GameState {
    StartScreen = 1,
    Playing = 2,
    EndScreen,
}
export enum TextColor { Highlighted='white', None='black' }
export enum TypedPart { None, Left, Middle }
export enum StoryKeys {
 testStory = 'testStory',
 testStory2 = 'testStory2',
 testStory3 = 'testStory3',
 threeLittlePigs = 'threeLittlePigs',
 frogOx = 'frogOx',
 bellingCat = 'bellingCat',
 townMouse = 'townMouse',
 foxGrapes = 'foxGrapes',
 wolfCrane = 'wolfCrane',
 lionMouse = 'lionMouse',
 gnatBull = 'gnatBull',
 owlGrasshopper = 'owlGrasshopper',
}